+++ stac_fastapi/extensions/stac_fastapi/extensions/sort/__init__.py
@@ -1,5 +1,17 @@
 """Sort extension module."""
 
-from .sort import SortConformanceClasses, SortExtension
+from .sort import (
+    CollectionSearchSortExtension,
+    ItemCollectionSortExtension,
+    SearchSortExtension,
+    SortConformanceClasses,
+    SortExtension,
+)
 
-__all__ = ["SortExtension", "SortConformanceClasses"]
+__all__ = [
+    "SortExtension",
+    "SearchSortExtension",
+    "ItemCollectionSortExtension",
+    "CollectionSearchSortExtension",
+    "SortConformanceClasses",
+]